Corrected error for graphs, the values need to be float.

+++ app/src/main/java/com/example/mainactivity/Activities/GraphStationActivity.java
@@ -144,7 +144,7 @@ public class GraphStationActivity extends AppCompatActivity {
                         }
                         else
                         {
-                            valor = Integer.parseInt(columnas[j]);
+                            valor = Float.parseFloat(columnas[j]);
                         }
                         break;
                     default:
